refactor(editor): organise content
