+++ src/layouts/components/Editor/components/MenuBar.tsx
@@ -5,9 +5,9 @@ import { v4 as uuid } from "uuid"
 
 import { useEditorModal } from "contexts/EditorModalContext"
 
-import MenuItem from "./MenuItem"
+import { MenuItem } from "./MenuItem"
 
-export default ({ editor }: { editor: Editor }) => {
+export const MenuBar = ({ editor }: { editor: Editor }) => {
   const { showModal } = useEditorModal()
 
   const items = [
