+++ src/layouts/components/Editor/extensions/Iframe/Iframe.ts
@@ -22,7 +22,7 @@ declare module "@tiptap/core" {
   }
 }
 
-export default Node.create<IframeOptions>({
+export const Iframe = Node.create<IframeOptions>({
   name: "iframe",
 
   group: "block",
