+++ src/layouts/components/Editor/extensions/Iframe/index.ts
@@ -1,0 +1,2 @@
+export * from "./Iframe"
+export * from "./IframeView"
