+++ src/layouts/components/Editor/extensions/index.ts
@@ -1,0 +1,1 @@
+export * from "./Iframe"
